+++ src/Button2/index.ts
@@ -1,9 +1,8 @@
-import {ButtonComponent} from './Button'
+import {Button as ButtonComponent} from './Button'
 import {Counter} from './ButtonCounter'
 import {IconButton} from './IconButton'
-import {LinkButton} from './LinkButton'
 export type {ButtonProps, IconButtonProps} from './types'
-export {IconButton, LinkButton}
+export {IconButton}
 
 export const Button = Object.assign(ButtonComponent, {
   Counter
